Feat/transactions

+++ app/models/user.py
@@ -8,6 +8,6 @@ users = sa.Table(
     sa.Column("name", sa.String(120), nullable=False),
     sa.Column("email", sa.String(120), nullable=False, unique=True),
     sa.Column("password", sa.String, nullable=False),
-    sa.Column("amount", sa.Integer, nullable=False, server_default="0"),
+    sa.Column("amount", sa.Numeric(12, 2), nullable=False, server_default="0"),
     sa.CheckConstraint("amount >= 0", name="check_users_amount_non_negative"),
 )
